Post alpha test changes, still need to make values save

+++ module/actors/sheets/character-sheet.js
@@ -226,8 +226,11 @@ export class STACharacterSheet extends ActorSheet {
         // Allows item-delete images to allow deletion of the selected item. This uses Simple Worldbuilding System Code.
         html.find('.control .delete').click(ev => {
             const li = $(ev.currentTarget).parents(".entry");
-            this.actor.deleteOwnedItem(li.data("itemId"));
-            li.slideUp(200, () => this.render(false));
+            var r = confirm("Are you sure you want to delete " + li[0].getAttribute("data-item-value") + "?");
+            if (r == true) {
+                this.actor.deleteOwnedItem(li.data("itemId"));
+                li.slideUp(200, () => this.render(false));
+            }
         });
 
         // Reads if a reputation track box has been clicked, and if it has will either: set the value to the clicked box, or reduce the value by one.
